complite projecye setup own

+++ calculator.c
@@ -5,17 +5,14 @@
 int main()
 {
     char expr[256];
-
-
     printf("Enter the mathematical expression: ");
     if (!fgets(expr, sizeof(expr), stdin))
     {
         printf("Input error\n");
         return 1;
     }
-
     token tokens[100];
-    int n = tokenize(expr, tokens, 100);
+    double n = tokenize(expr, tokens, 100);
     if (n < 0)
     {
         printf("Failed to parse the expression\n");
@@ -32,9 +29,7 @@ int main()
             printf("operation: '%c'\n", (char)tokens[i].type);
         }
     }
-
     Word result = evaluate(tokens, n);
     printf("Output : %lld\n", (long long)result);
-
     return 0;
 }
